Fix some ALFE build and warning issues

+++ include/alfe/hash.h
@@ -9,7 +9,10 @@ class Hash
 {
 public:
     // FNV-1a hash
-    Hash(const std::type_info& t) : _h(0x811c9dc5) { mixin(t.hash_code()); }
+    Hash(const std::type_info& t) : _h(0x811c9dc5)
+    {
+        mixin(static_cast<UInt32>(t.hash_code()));
+    }
     Hash& mixin(UInt32 v) { _h = (_h ^ v) * 0x01000193; return *this; }
     operator UInt32() const { return _h; }
 private:
